Refactor application files to delete unnecessary files and follow style guidelines
