--- src/Components/Description.css
@@ -1,10 +1,0 @@
-.image-description {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 30vw;
-  height: 50vh;
-  margin: 3em;
-  font-size: 1.2em;
-}
