+++ src/Components/Error.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import './Error.css';
 import { Link } from 'react-router-dom';
+import '../styling/Error.css';
 
 const Error = () => {
   return (
@@ -9,7 +9,7 @@ const Error = () => {
       <Link to='/' className='redirect'>Click here to go back to the main gallery</Link>
       <img className="fof-image" src="https://i.kym-cdn.com/photos/images/original/000/548/129/538.jpg" />
     </div>
-  )
-}
+  );
+};
 
 export default Error;
