--- src/Components/Frame.css
@@ -1,5 +1,0 @@
-.art-frame {
-  width: 50vw;
-  max-height: 50vh;
-  margin: 3em;
-}
