+++ src/Components/Frame.js
@@ -1,12 +1,13 @@
-import React from 'react'
-import Art from './Art'
-import './Frame.css'
+import React from 'react';
+import Art from './Art';
+import '../styling/Frame.css';
 
 const Frame = () => {
   return (
     <div className="art-frame">
       <Art />
     </div>
-  )
-}
-export default Frame
+  );
+};
+
+export default Frame;
