--- src/Components/Nav.css
@@ -1,34 +1,0 @@
-nav {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.page-title {
-  font-size: 4em;
-}
-
-.btn-styling {
-  display: flex;
-  width: 50%;
-  justify-content: space-around;
-}
-
-button {
-  width: 12em;
-  font-size: 1.2em;
-  height: 3em;
-  border: none;
-  margin: 0.6em;
-  border-radius: 0.2em;
-  box-shadow: 0px 4px 8px #AA771C;
-  background: radial-gradient(ellipse farthest-corner at right bottom, #BF953F, #FCF6BA 100%);
-  cursor: pointer;
-  font-family: 'Special Elite', cursive;
-}
-
-button:hover {
-  transform: scale(1.06);
-  transition: .6s;
-}
