+++ src/index.js
@@ -4,8 +4,7 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from './app/store';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
-import './index.css';
+import './styling/index.css';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
@@ -17,8 +16,3 @@ root.render(
       </BrowserRouter>
     </Provider>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
