+++ src/styling/App.css
@@ -1,0 +1,4 @@
+.App {
+  font-family: "Special Elite", cursive;
+  text-align: center;
+}
