+++ src/styling/Error.css
@@ -1,17 +1,17 @@
 .fof-box {
-  display: flex;
-  flex-direction: column;
   font-size: 2em;
-  justify-content: center;
+  display: flex;
   align-items: center;
+  flex-direction: column;
+  justify-content: center;
 }
 
 .fof-image {
-  height: 18em;
-  width: 10em;
   border: solid #bf953f 7px;
   border-style: outset;
+  height: 18em;
   padding: 1em;
+  width: 10em;
 }
 
 .redirect {
