+++ src/styling/Frame.css
@@ -1,0 +1,5 @@
+.art-frame {
+  margin: 3em;
+  max-height: 50vh;
+  width: 50vw;
+}
