+++ src/styling/index.css
@@ -1,0 +1,5 @@
+body {
+  background-color: rgb(211, 242, 247);
+  height: 100vh;
+  margin: 0;
+}
